WS-009: Rename client file
